Delete permissions.ts
